# __docs__/api/openapi.yaml
openapi: 3.1.0
info:
  title: Agonez Atlas API
  description: REST API for the Agonez exercise and muscle Atlas and the relational
    PlanCreator draft editor and the desktop Execution workflow API. Authentication
    and ownership are intentionally deferred.
  version: 0.1.0
paths:
...
  /api/v1/exec/plan-runs:
    get:
      tags:
      - Execution
      summary: List Plan Runs
      operationId: list_plan_runs_api_v1_exec_plan_runs_get
      parameters:
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - type: array
            items:
              $ref: '#/components/schemas/PlanRunStatus'
          - type: 'null'
          title: Status
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          default: 50
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlanRunListResponse'
        '404':
          description: Execution resource not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '409':
          description: Execution state conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422':
          description: Execution request validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
    post:
      tags:
      - Execution
      summary: Create Plan Run
      operationId: create_plan_run_api_v1_exec_plan_runs_post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlanRunCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlanRunOverview'
        '404':
          description: Execution resource not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '409':
          description: Execution state conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422':
          description: Execution request validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/v1/exec/plan-runs/preview:
    get:
      tags:
      - Execution
      summary: Preview Plan Run
      operationId: preview_plan_run_api_v1_exec_plan_runs_preview_get
      parameters:
      - name: plan_revision_id
        in: query
        required: true
        schema:
          type: integer
          minimum: 1
          title: Plan Revision Id
      - name: starts_on
        in: query
        required: true
        schema:
          type: string
          format: date
          title: Starts On
      - name: microcycle_count
        in: query
        required: true
        schema:
          type: integer
          title: Microcycle Count
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlanRunPreview'
        '404':
          description: Execution resource not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '409':
          description: Execution state conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422':
          description: Execution request validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/v1/exec/plan-runs/{plan_run_id}/overview:
    get:
      tags:
      - Execution
      summary: Get Plan Run Overview
      operationId: get_plan_run_overview_api_v1_exec_plan_runs__plan_run_id__overview_get
      parameters:
      - name: plan_run_id
        in: path
        required: true
        schema:
          type: integer
          minimum: 1
          title: Plan Run Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlanRunOverview'
        '404':
          description: Execution resource not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '409':
          description: Execution state conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422':
          description: Execution request validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/v1/exec/plan-runs/{plan_run_id}:
    patch:
      tags:
      - Execution
      summary: Update Plan Run
      operationId: update_plan_run_api_v1_exec_plan_runs__plan_run_id__patch
      parameters:
      - name: plan_run_id
        in: path
        required: true
        schema:
          type: integer
          minimum: 1
          title: Plan Run Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlanRunPatch'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlanRunOverview'
        '404':
          description: Execution resource not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '409':
          description: Execution state conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422':
          description: Execution request validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/v1/exec/plan-runs/{plan_run_id}/analysis/queue:
    get:
      tags:
      - Execution
      summary: Get Analysis Queue
      operationId: get_analysis_queue_api_v1_exec_plan_runs__plan_run_id__analysis_queue_get
      parameters:
      - name: plan_run_id
        in: path
        required: true
        schema:
          type: integer
          minimum: 1
          title: Plan Run Id
      - name: target_microcycle
        in: query
        required: false
        schema:
          anyOf:
          - type: integer
            minimum: 1
          - type: 'null'
          title: Target Microcycle
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalysisQueue'
        '404':
          description: Execution resource not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '409':
          description: Execution state conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422':
          description: Execution request validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/v1/exec/plan-runs/{plan_run_id}/exercise-traces/{exercise_trace_id}:
    get:
      tags:
      - Execution
      summary: Get Exercise Trace
      operationId: get_exercise_trace_api_v1_exec_plan_runs__plan_run_id__exercise_traces__exercise_trace_id__get
      parameters:
      - name: plan_run_id
        in: path
        required: true
        schema:
          type: integer
          minimum: 1
          title: Plan Run Id
      - name: exercise_trace_id
        in: path
        required: true
        schema:
          type: integer
          minimum: 1
          title: Exercise Trace Id
      - name: include_next
        in: query
        required: false
        schema:
          type: boolean
          default: true
          title: Include Next
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExerciseTraceResponse'
        '404':
          description: Execution resource not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '409':
          description: Execution state conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422':
          description: Execution request validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/v1/exec/plan-runs/{plan_run_id}/exercise-traces/{exercise_trace_id}/next:
    get:
      tags:
      - Execution
      summary: Get Next Prescription
      operationId: get_next_prescription_api_v1_exec_plan_runs__plan_run_id__exercise_traces__exercise_trace_id__next_get
      parameters:
      - name: plan_run_id
        in: path
        required: true
        schema:
          type: integer
          minimum: 1
          title: Plan Run Id
      - name: exercise_trace_id
        in: path
        required: true
        schema:
          type: integer
          minimum: 1
          title: Exercise Trace Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NextPrescription'
        '404':
          description: Execution resource not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '409':
          description: Execution state conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422':
          description: Execution request validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/v1/exec/plan-runs/{plan_run_id}/sessions/{session_id}/exercise-prescriptions/{exercise_trace_id}:
    put:
      tags:
      - Execution
      summary: Save Exercise Prescription
      operationId: save_exercise_prescription_api_v1_exec_plan_runs__plan_run_id__sessions__session_id__exercise_prescriptions__exercise_trace_id__put
      parameters:
      - name: plan_run_id
        in: path
        required: true
        schema:
          type: integer
          minimum: 1
          title: Plan Run Id
      - name: session_id
        in: path
        required: true
        schema:
          type: integer
          minimum: 1
          title: Session Id
      - name: exercise_trace_id
        in: path
        required: true
        schema:
          type: integer
          minimum: 1
          title: Exercise Trace Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExercisePrescriptionPut'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PrescriptionWriteResponse'
        '404':
          description: Execution resource not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '409':
          description: Execution state conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422':
          description: Execution request validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
    delete:
      tags:
      - Execution
      summary: Delete Exercise Prescription
      operationId: delete_exercise_prescription_api_v1_exec_plan_runs__plan_run_id__sessions__session_id__exercise_prescriptions__exercise_trace_id__delete
      parameters:
      - name: plan_run_id
        in: path
        required: true
        schema:
          type: integer
          minimum: 1
          title: Plan Run Id
      - name: session_id
        in: path
        required: true
        schema:
          type: integer
          minimum: 1
          title: Session Id
      - name: exercise_trace_id
        in: path
        required: true
        schema:
          type: integer
          minimum: 1
          title: Exercise Trace Id
      responses:
        '204':
          description: Successful Response
        '404':
          description: Execution resource not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '409':
          description: Execution state conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422':
          description: Execution request validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/v1/exec/plan-runs/{plan_run_id}/sessions/{session_id}/prescription:
    patch:
      tags:
      - Execution
      summary: Update Workout Prescription
      operationId: update_workout_prescription_api_v1_exec_plan_runs__plan_run_id__sessions__session_id__prescription_patch
      parameters:
      - name: plan_run_id
        in: path
        required: true
        schema:
          type: integer
          minimum: 1
          title: Plan Run Id
      - name: session_id
        in: path
        required: true
        schema:
          type: integer
          minimum: 1
          title: Session Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkoutPrescriptionPatch'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkoutPrescriptionResponse'
        '404':
          description: Execution resource not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '409':
          description: Execution state conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422':
          description: Execution request validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/v1/exec/plan-runs/{plan_run_id}/workout-traces:
    get:
      tags:
      - Execution
      summary: List Workout Traces
      operationId: list_workout_traces_api_v1_exec_plan_runs__plan_run_id__workout_traces_get
      parameters:
      - name: plan_run_id
        in: path
        required: true
        schema:
          type: integer
          minimum: 1
          title: Plan Run Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkoutTraceList'
        '404':
          description: Execution resource not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '409':
          description: Execution state conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422':
          description: Execution request validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/v1/exec/plan-runs/{plan_run_id}/workout-traces/{workout_trace_id}:
    get:
      tags:
      - Execution
      summary: Get Workout Trace
      operationId: get_workout_trace_api_v1_exec_plan_runs__plan_run_id__workout_traces__workout_trace_id__get
      parameters:
      - name: plan_run_id
        in: path
        required: true
        schema:
          type: integer
          minimum: 1
          title: Plan Run Id
      - name: workout_trace_id
        in: path
        required: true
        schema:
          type: integer
          minimum: 1
          title: Workout Trace Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkoutTraceResponse'
        '404':
          description: Execution resource not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '409':
          description: Execution state conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422':
          description: Execution request validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/v1/exec/plan-runs/{plan_run_id}/microcycles:
    get:
      tags:
      - Execution
      summary: Get Microcycles
      operationId: get_microcycles_api_v1_exec_plan_runs__plan_run_id__microcycles_get
      parameters:
      - name: plan_run_id
        in: path
        required: true
        schema:
          type: integer
          minimum: 1
          title: Plan Run Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MicrocycleTimeline'
        '404':
          description: Execution resource not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '409':
          description: Execution state conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422':
          description: Execution request validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/v1/exec/plan-runs/{plan_run_id}/microcycles/{ordinal}:
    patch:
      tags:
      - Execution
      summary: Update Microcycle
      operationId: update_microcycle_api_v1_exec_plan_runs__plan_run_id__microcycles__ordinal__patch
      parameters:
      - name: plan_run_id
        in: path
        required: true
        schema:
          type: integer
          minimum: 1
          title: Plan Run Id
      - name: ordinal
        in: path
        required: true
        schema:
          type: integer
          minimum: 1
          title: Ordinal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MicrocyclePatch'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TimelineMicrocycle'
        '404':
          description: Execution resource not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '409':
          description: Execution state conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422':
          description: Execution request validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
components:
  schemas:
...
    ErrorEnvelope:
      properties:
        error:
          $ref: '#/components/schemas/ErrorBody'
      additionalProperties: false
      type: object
      required:
      - error
      title: ErrorEnvelope
...
    PlanRunCreate:
      properties:
        plan_revision_id:
          type: integer
          minimum: 1.0
          title: Plan Revision Id
        name:
          type: string
          maxLength: 200
          minLength: 1
          title: Name
        starts_on:
          type: string
          format: date
          title: Starts On
        microcycle_count:
          type: integer
          title: Microcycle Count
      additionalProperties: false
      type: object
      required:
      - plan_revision_id
      - name
      - starts_on
      - microcycle_count
      title: PlanRunCreate
...
    PlanRunListResponse:
      properties:
        as_of:
          type: string
          format: date
          title: As Of
        items:
          items:
            $ref: '#/components/schemas/PlanRunListItem'
          type: array
          title: Items
      additionalProperties: false
      type: object
      required:
      - as_of
      - items
      title: PlanRunListResponse
...
    PlanRunPatch:
      properties:
        name:
          anyOf:
          - type: string
            maxLength: 200
            minLength: 1
          - type: 'null'
          title: Name
        status:
          anyOf:
          - $ref: '#/components/schemas/PlanRunStatus'
          - type: 'null'
        expected_version:
          type: string
          title: Expected Version
      additionalProperties: false
      type: object
      required:
      - expected_version
      title: PlanRunPatch
...
    PlanRunStatus:
      type: string
      enum:
      - scheduled
      - active
      - cancelled
      - completed
      title: PlanRunStatus
...